refactor: 팔로우 및 언팔로우 동시성 처리 및 테스트 코드 구현
